Fixes https://github.com/canonical/lxd/issues/17159

+++ test/includes/lxd.sh
@@ -233,12 +233,14 @@ kill_lxd() {
                 timeout -k 20 20 lxc storage volume delete "${storage_pool}" "${volume}" --force-local || true
             done < <(lxc query "${path}" | jq --exit-status --raw-output ".[] | ltrimstr(\"${path}/\")")
 
-            # Delete the storage buckets.
-            path="/1.0/storage-pools/${storage_pool}/buckets"
-            while read -r bucket; do
-                echo "   ⛔ Deleting storage bucket ${bucket} on ${storage_pool}"
-                timeout -k 20 20 lxc storage bucket delete "${storage_pool}" "${bucket}" --force-local || true
-            done < <(lxc query "${path}" | jq --exit-status --raw-output ".[] | ltrimstr(\"${path}/\")")
+            # Delete the storage buckets (only object storage pools support buckets).
+            if lxc query "/1.0/storage-pools/${storage_pool}" | jq --exit-status '.driver == "cephobject"' > /dev/null; then
+                path="/1.0/storage-pools/${storage_pool}/buckets"
+                while read -r bucket; do
+                    echo "   ⛔ Deleting storage bucket ${bucket} on ${storage_pool}"
+                    timeout -k 20 20 lxc storage bucket delete "${storage_pool}" "${bucket}" --force-local || true
+                done < <(lxc query "${path}" | jq --exit-status --raw-output ".[] | ltrimstr(\"${path}/\")")
+            fi
 
             ## Delete the storage pool.
             echo "   ⛔ Deleting storage pool ${storage_pool}"
